refactor: change dpi type annotation

+++ src/qaa/commands/cmd_pca.py
@@ -131,7 +131,7 @@ def cli(
     whiten: bool,
     bias: bool,
     image: bool,
-    dpi: float,
+    dpi: int,
     image_type: str,
     verbose: bool,
 ):
